requirements.txt
aiogram==3.7.0
openai==1.34.0
pydub==0.25.1
pydantic==2.7.4
python-dotenv==1.0.1
SQLAlchemy==2.0.31
alembic==1.13.1